Move service configuration classes outside simulation runner module
