+++ cstar/entrypoint/service.py
@@ -1,5 +1,4 @@
 import asyncio
-import logging
 import signal
 import time
 from abc import ABC, abstractmethod
@@ -7,49 +6,12 @@ from queue import Empty, Full, Queue
 from threading import Event, Thread
 from typing import TYPE_CHECKING, ClassVar, Final, Literal
 
-from pydantic import BaseModel, Field
-
 from cstar.base.log import LoggingMixin
 
 if TYPE_CHECKING:
     from types import FrameType
 
-
-class ServiceConfiguration(BaseModel):
-    """Configuration options for a Service."""
-
-    as_service: bool = False
-    """Determines lifetime of the service.
-
-    When `True`, calling `execute` on the service will run continuously until
-    shutdown criteria are met. When `False`, the service completes a single
-    pass through the service lifecycle and automatically exits.
-    """
-    loop_delay: float = Field(default=0.0, ge=0.0)
-    """Duration (in seconds) of a delay between iterations of the main event loop."""
-    health_check_frequency: float | None = Field(default=None, ge=0.0)
-    """Time (in seconds) between calls to a health check handler.
-
-    NOTE:
-    - A value of `None` disables health checks.
-    - A value of `0` triggers the health check on every iteration.
-    """
-    log_level: int = logging.INFO
-    """The logging level used by the service."""
-    health_check_log_threshold: int = Field(default=10, ge=3)
-    """The number of health-checks that may be missed before logging."""
-    name: str = "Service"
-    """A user-friendly name for logging."""
-
-    @property
-    def healthcheck_enabled(self) -> bool:
-        """Return `True` when the health check frequency is non-null.
-
-        Returns
-        -------
-        bool
-        """
-        return self.health_check_frequency is not None
+    from cstar.entrypoint.config import ServiceConfiguration
 
 
 class Service(ABC, LoggingMixin):
@@ -62,7 +24,7 @@ class Service(ABC, LoggingMixin):
     CMD_HEARTBEAT: Literal["heartbeat"] = "heartbeat"
     MIN_HCF: ClassVar[float] = 0.01
 
-    _config: Final[ServiceConfiguration]
+    _config: Final["ServiceConfiguration"]
     """Runtime configuration of the `Service`"""
     _hc_thread: Thread | None = None
     """A thread for executing an unblocked healthcheck callback."""
@@ -73,7 +35,7 @@ class Service(ABC, LoggingMixin):
 
     def __init__(
         self,
-        config: ServiceConfiguration,
+        config: "ServiceConfiguration",
     ) -> None:
         """Initialize the Service.
 
@@ -232,7 +194,7 @@ class Service(ABC, LoggingMixin):
 
     def _healthcheck(
         self,
-        config: ServiceConfiguration,
+        config: "ServiceConfiguration",
         msg_queue: Queue,
     ) -> None:
         """Run the health-check event loop.
